requirements.txt
langchain==0.3.27
langchain-google-genai==2.0.10
langchain-ibm==0.3.19
langchain-text-splitters
langchain-community
pypdf
python-dotenv
Flask
gunicorn
werkzeug
